[tests] Moved test plugin folder
