+++ .drone.yml
@@ -12,7 +12,7 @@ steps:
     commands:
       - pip3 install -r requirements.txt --user
       - pip3 install -r test-requirements.txt --user
-      - python3 -m pytest -p pycti.connector.tests.pytest_pycti --no-header -vv --disable-warnings --cov=pycti --drone --connectors
+      - python3 -m pytest -p pycti.test_plugin.pytest_pycti --no-header -vv --disable-warnings --cov=pycti --drone --connectors
   - name: client-test-39
     image: python:3.9
     commands:
