+++ pycti/test_plugin/utils.py
