+++ setup.cfg
@@ -31,9 +31,9 @@ packages =
     pycti.connector.libs
     pycti.connector.libs.messaging
     pycti.connector.libs.mixins
-    pycti.connector.tests
     pycti.connector.connector_types
     pycti.entities
+    pycti.test_plugin
     pycti.utils
 include_package_data = True
 install_requires =
